pyproject.toml
[project]
name = "learn-a2a"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[all,encryption,grpc,http-server,telemetry]>=0.3.22",
    "fastapi>=0.128.0",
    "pydantic>=2.12.5",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
]
